Fixed build errors on modern Linux
Fixed the two build errors that happen on Linux with the previous
includes.

Added the 'cstring'/'string.h' include for the 'always.h' header to
include the C string functions.

Added the 'unicode/ustring.h' include for the 'unichar.h' header to
include the unicode string utilities.

The MSVC compiler may be able to find these methods independently but
GCC requires the explicit inclusion of the header files.

Signed-off-by: Victor Matia <vmatir@gmail.com>